Facet の結果部分を /proteins/* を辞め、/report_type/* にした
