--- app/controllers/proteins/environment_controller.rb
@@ -1,2 +1,0 @@
-class Proteins::EnvironmentController < ProteinsController
-end
