--- app/controllers/proteins/gene_controller.rb
@@ -1,2 +1,0 @@
-class Proteins::GeneController < ProteinsController
-end
