--- app/controllers/proteins/organism_controller.rb
@@ -1,2 +1,0 @@
-class Proteins::OrganismController < ProteinsController
-end
