--- app/controllers/proteins/phenotype_controller.rb
@@ -1,2 +1,0 @@
-class Proteins::PhenotypeController < ApplicationController
-end
