+++ app/controllers/report_type/environments_controller.rb
@@ -1,0 +1,2 @@
+class ReportType::EnvironmentsController < ReportType::BaseController
+end
