+++ app/controllers/report_type/genes_controller.rb
@@ -1,0 +1,2 @@
+class ReportType::GenesController < ReportType::BaseController
+end
