+++ app/controllers/report_type/organisms_controller.rb
@@ -1,0 +1,2 @@
+class ReportType::OrganismsController < ReportType::BaseController
+end
