+++ app/controllers/report_type/phenotypes_controller.rb
@@ -1,0 +1,2 @@
+class ReportType::PhenotypesController < ApplicationController
+end
