+++ app/controllers/report_type/proteins_controller.rb
@@ -1,0 +1,2 @@
+class ReportType::ProteinsController < ReportType::BaseController
+end
