--- app/models/proteins/gene.rb
@@ -1,6 +1,0 @@
-# coding: utf-8
-
-module Proteins
-  class Gene < ::Protein
-  end
-end
