+++ app/models/report_type/gene.rb
@@ -1,0 +1,4 @@
+module ReportType
+  class Gene < Base
+  end
+end
