+++ app/models/report_type/organism.rb
@@ -1,5 +1,5 @@
-module Proteins
-  class Organism < ::Protein
+module ReportType
+  class Organism < Base
     class << self
       def count(meo_id: '', tax_id: '', bp_id: '', mf_id: '', cc_id: '', mpo_id: '')
         sparql  = organism_count_sparql(meo_id, tax_id, bp_id, mf_id, cc_id, mpo_id)
@@ -40,9 +40,9 @@ module Proteins
     end
 
     def initialize(up_tax, envs, phenotypes)
-      @tax = ::Taxonomy.new(up_tax[:taxonomy_id], up_tax[:taxonomy_name])
-      @envs       = envs.map {|env| ::Environment.new(env[:meo_id], env[:meo_name]) }
-      @phenotypes = phenotypes.map {|phenotype| ::Phenotype.new(phenotype[:mpo_id], phenotype[:mpo_name]) }
+      @tax        = Base::Taxonomy.new(up_tax[:taxonomy_id], up_tax[:taxonomy_name])
+      @envs       = envs.map {|env| Base::Environment.new(env[:meo_id], env[:meo_name]) }
+      @phenotypes = phenotypes.map {|phenotype| Base::Phenotype.new(phenotype[:mpo_id], phenotype[:mpo_name]) }
     end
 
     attr_reader :tax, :envs, :phenotypes
