+++ app/models/report_type/protein.rb
@@ -1,0 +1,4 @@
+module ReportType
+  class Protein < Base
+  end
+end
