+++ app/views/report_type/environments/_results.html.haml
@@ -1,5 +1,5 @@
 = content_for :javascript do
-  = javascript_include_tag 'environment_results_table'
+  = javascript_include_tag 'results_table/environment'
 
 %table#environment_results.table.table-striped.table-bordered.table-hover.table-condensed
   %thead
