+++ app/views/report_type/environments/search.json.jbuilder
