+++ app/views/report_type/genes/_results.html.haml
@@ -1,5 +1,5 @@
 = content_for :javascript do
-  = javascript_include_tag 'gene_results_table'
+  = javascript_include_tag 'results_table/gene'
 
 %table#gene_results.table.table-striped.table-bordered.table-hover.table-condensed
   %thead
