+++ app/views/report_type/genes/search.csv.csvbuilder
@@ -10,17 +10,11 @@ csv << [
   'Organism',
   'Organism ID',
   'Organism URL',
-  'Environments',
-  'Environments ID',
-  'Environments URL',
-  'Phenotypes',
-  'Phenotypes ID',
-  'Phenotypes URL'
 ]
 
 limit = 100
 (0..@hits_count).step(limit) do |offset|
-  Protein.search(@args.merge(limit: limit, offset:offset)).each do |p|
+  ReportType::Gene.search(@args.merge(limit: limit, offset:offset)).each do |p|
     csv << [
       p.name,
       p.genes.map(&:id).join("\n"),
@@ -33,12 +27,6 @@ limit = 100
       p.tax.name,
       p.tax.id,
       "http://togogenome.org/organism/#{p.tax.id}",
-      p.envs.map(&:name).join("\n"),
-      p.envs.map(&:id).join("\n"),
-      p.envs.map {|env| "http://togogenome.org/environment/#{env.id}" }.join("\n"),
-      p.phenotypes.map(&:name).join("\n"),
-      p.phenotypes.map(&:id).join("\n"),
-      p.phenotypes.map {|phenotype| "http://togogenome.org/phenotype/#{phenotype.id}" }.join("\n")
     ]
   end
 end
