+++ app/views/report_type/genes/search.json.jbuilder
