+++ app/views/report_type/organisms/_results.html.haml
@@ -1,5 +1,5 @@
 = content_for :javascript do
-  = javascript_include_tag 'organism_results_table'
+  = javascript_include_tag 'results_table/organism'
 
 %table#organism_results.table.table-striped.table-bordered.table-hover.table-condensed
   %thead
