+++ app/views/report_type/organisms/search.csv.csvbuilder
@@ -12,7 +12,7 @@ csv << [
 
 limit = 100
 (0..@hits_count).step(limit) do |offset|
-  Proteins::Organism.search(@args.merge(limit: limit, offset:offset)).each do |p|
+  ReportType::Organism.search(@args.merge(limit: limit, offset:offset)).each do |p|
     csv << [
       p.tax.name,
       p.tax.id,
