+++ app/views/report_type/organisms/search.json.jbuilder
