+++ app/views/report_type/phenotypes/_results.html.haml
