+++ app/views/report_type/proteins/_results.html.haml
@@ -1,7 +1,7 @@
 = content_for :javascript do
-  = javascript_include_tag 'results_table'
+  = javascript_include_tag 'results_table/protein'
 
-%table#results.table.table-striped.table-bordered.table-hover.table-condensed
+%table#protein_results.table.table-striped.table-bordered.table-hover.table-condensed
   %thead
     %td Description
     %td Gene
