+++ app/views/report_type/proteins/search.json.jbuilder
