+++ config/environments/production.rb
@@ -48,8 +48,7 @@ TogoGenome::Application.configure do
   # config.action_controller.asset_host = "http://assets.example.com"
 
   # Precompile additional assets (application.js, application.css, and all non-JS/CSS are already added)
-  config.assets.precompile += %w(togo.css togogenome.js d3.js stanza_search.js facet_results.js
-    results_table.js gene_results_table.js organism_results_table.js environment_results_table.js)
+  config.assets.precompile += %w(togo.css togogenome.js d3.js stanza_search.js facet_results.js results_table/*)
 
   # Disable delivery errors, bad email addresses will be ignored
   # config.action_mailer.raise_delivery_errors = false
