+++ config/routes.rb
@@ -5,14 +5,12 @@ TogoGenome::Application.routes.draw do
     get :search, on: :member
   end
 
-  get '/proteins/search', as: :search_proteins
-  get '/proteins/all/search', to: 'proteins#search'
-
-  namespace :proteins do
-    get '/gene/search'
-    get '/organism/search'
-    get '/environment/search'
-    get '/phenotype/search'
+  namespace :report_type do
+    get '/proteins/search'
+    get '/genes/search'
+    get '/organisms/search'
+    get '/environments/search'
+    get '/phenotypes/search'
   end
 
   resources :genomes, only: %w(index) do
